Performing full refresh in scenarios where client/server are out of sync (on transaction id, i.e., scenario where two browsers are connected to same session.

+++ src/server-java/webcontainer/nextapp/echo/webcontainer/ClientMessage.java
@@ -73,7 +73,7 @@ public class ClientMessage {
         return document;
     }
     
-    public long getTransactionId() {
+    public int getTransactionId() {
         return transactionId;
     }
